+++ changelog.md
@@ -1,7 +1,23 @@
 # Changelog
 
 **./public/scripts/market-data.js**
-* When toggling table rows, the 'display' property is now used instead of 'visibility'
-	* Prevents hidden rows from creating unnecessary whitespace.
-	* There is a proper way to display table row elements.
-	* Related to the 'callSearch' and 'renderCurrencyData' functions.
+* Moved table definition in the 'initializeMarketTable' function to 'index.ejs'
+	* HTML is static and not dynamic.
+* Removed the following from 'initializeMarketTable'
+	* 'tableDispElement', 'renderedTable' variables.
+	* 'setHeaderRow', 'setTableBody' calls.
+	* 'searchTextbox' variable - Textbox value is set directly.
+* Removed functions:
+	* setHeaderRow
+	* setTableBody
+	* addColumn
+* Changes to 'handleDataSort'
+	* Removed 'clickEvent' parameter, the click event itself.
+	* Added 'sortButton' parameter, the button being clicked.
+	* Replaced 'clickEvent.target' with 'sortButton'
+
+---
+
+**./views/index.ejs**
+* Data table skeleton is defined here as static HTML
+* Moved from '/scripts/market-data.js'